+++ src/components/home/hero/HeroGoalSection.tsx
